+++ DataLAE_v0.2/src/Service/LocalService.java
@@ -37,11 +37,24 @@ public final class LocalService implements ILocalService {
 
 	@Override
 	public void inserir(Local l) throws SQLException {
+		if (l == null)
+			throw new IllegalArgumentException("Argumento nulo!"); 
+		
+		if (l.getNome() == null)
+			throw new IllegalArgumentException("Campo obrigatório nulo!"); 
+		
 		dao.inserir(l);
 	}
 
 	@Override
 	public void remover(Local l) throws SQLException {
+		if (l == null)
+			throw new IllegalArgumentException("Argumento nulo!"); 
+		
+		if (l.getCodigo() == null)
+			throw new IllegalArgumentException("Impossível encontrar o local informado"); 
+		
+		
 		dao.remover(l);
 	}
 
@@ -52,11 +65,24 @@ public final class LocalService implements ILocalService {
 
 	@Override
 	public ArrayList<Local> buscar(Local l, Pesquisador p, Projeto proj) throws SQLException {
+		if (l == null)
+			throw new IllegalArgumentException("Argumento nulo!"); 
+		
+		if (l.getCodigo() == null)
+			throw new IllegalArgumentException("Impossível encontrar o local informado"); 
+		
 		return dao.buscar(l, p, proj);
 	}
 
 	@Override
 	public void alterar(Local l) {
+		if (l == null)
+			throw new IllegalArgumentException("Argumento nulo!"); 
+		
+		if (l.getCodigo() == null)
+			throw new IllegalArgumentException("Impossível encontrar o local informado"); 
+		
+		
 		dao.alterar(l);
 	}
